[refac] descriptor: move Descriptor Key logic to key submodule
